# msg/async/dpdk: Use hash map to manage mbuf releasing