@rosinL 2020-03-26
jenkins test make check